# http5/homework/chat-server/internal/handler/request/send_public_message.go
package request

import "github.com/go-playground/validator/v10"

type SendPublicMessageRequest struct {
	FromID  int    `json:"from_id" validate:"required,min=1"`
	Content string `json:"content" validate:"required,min=1,max=2000"`
}

func (sm *SendPublicMessageRequest) Validate(valid *validator.Validate) error {
	return valid.Struct(sm)
}